+++ app/serializers/daggerheart/homebrew/book_serializer.rb
@@ -8,8 +8,9 @@ module Daggerheart
       def items
         object_items = object.items.group_by(&:itemable_type).transform_values { |item| item.pluck(:itemable_id) }
         {
-          races: object_items['Daggerheart::Homebrew::Race'],
-          communities: object_items['Daggerheart::Homebrew::Community']
+          races: object_items['Daggerheart::Homebrew::Race'] || [],
+          communities: object_items['Daggerheart::Homebrew::Community'] || [],
+          subclasses: object_items['Daggerheart::Homebrew::Subclass'] || []
         }
       end
     end