+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - managing homebrew communities for Daggerheart
 - adding homebrew communities to modules
+- adding homebrew subclasses to modules
 
 ### Modified
 - errors rendering